requirements.txt
openai-whisper>=20240124
ffmpeg-python>=0.2.0
torch>=2.0.0
numpy>=1.20.0
tqdm>=4.65.0
flask>=3.0.0
werkzeug>=3.0.0 